Fix PLL int64 overflow, protect IN channel, remove dead code
- pll.c: e_norm (int64 << 15), err sum in int64, ki*err in int64
  (prevents crash at high speed / large EMF)
- protect.c: overcurrent checks I1 + I2 + IN (all three channels)
- pwm.c: remove dead dtg variable in PWM_Init

+++ src/pll.c
@@ -25,20 +25,21 @@ void PLL_Update(PLL *pll, int32_t emf_alpha, int32_t emf_beta) {
     CORDIC_Modulus(emf_alpha, emf_beta, &mod, &angle);
     if(mod == 0) return;
 
-    int32_t e_norm_a = (emf_alpha << 15) / mod;
-    int32_t e_norm_b = (emf_beta  << 15) / mod;
+    int32_t e_norm_a = (int32_t)(((int64_t)emf_alpha << 15) / mod);
+    int32_t e_norm_b = (int32_t)(((int64_t)emf_beta  << 15) / mod);
 
-    /* Ошибка PLL: err = -Eα*sin(θ) + Eβ*cos(θ) */
+    /* Ошибка PLL: err = -Eα*sin(θ) + Eβ*cos(θ)
+     * Q15×Q15 = Q30; сумма двух Q30 может превысить int32 — считаем в int64 */
     int32_t s, c;
     CORDIC_SinCos((int32_t)pll->theta_u32, &s, &c);
-    int32_t err = (-e_norm_a * s + e_norm_b * c) >> 15;
+    int32_t err = (int32_t)((-(int64_t)e_norm_a * s + (int64_t)e_norm_b * c) >> 15);
 
     /* PI. Единицы omega: приращение угла q31 за один FOC-цикл (Ts).
      * Прямое интегрирование theta += omega без умножения на ts_us —
      * исключает переполнение int32 на высоких скоростях. */
     /* Anti-windup: ограничение интегратора (предотвращает раскрутку
      * при потере захвата / шуме EMF на малых оборотах). */
-    pll->integrator += (pll->ki * err) >> 15;
+    pll->integrator += (int32_t)(((int64_t)pll->ki * err) >> 15);
     if(pll->integrator > PLL_INTEGRATOR_MAX) pll->integrator = PLL_INTEGRATOR_MAX;
     if(pll->integrator < -PLL_INTEGRATOR_MAX) pll->integrator = -PLL_INTEGRATOR_MAX;
     int32_t omega = ((pll->kp * err) >> 15) + pll->integrator;
